refactor: 도메인 패키지 변경 및 에러코드 병합
